third_party/blink/web_tests/wpt_internal/display-lock/css-subtree-visibility/subtree-visibility-051.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Subtree Visibility: accessibility focus</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">
<meta name="assert" content="subtree-visibility auto subtrees are exposed by accessibility focus">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div style="height:10000px;">
  spacer so that everything below will be offscreen (and won't get viewport-activated)
</div>
<div id="hidden" style="subtree-visibility: auto">
  foo
  <div id="child" tabindex="0">
    bar
  </div>
</div>

<script>
function axElementById(id) {
  return accessibilityController.accessibleElementById(id);
}

async_test(async(t) => {
  const hiddenEl = document.getElementById("hidden");
  let axHidden = axElementById("hidden");
  // #hidden is locked, and thus is saved as an AXNodeObject instead of an AXLayoutObject,
  // and has 3 child nodes: "foo" text, "<newline>" text, and #child node.
  // Note that we have the separate <newline> text because it's an AXNodeObject and does
  // not do whitespace collapsing (though this might change in the future).
  t.step(() => { assert_equals(axHidden.childrenCount, 3, "Children count when locked"); });

  // Focus to #child, causing #hidden to get unlocked.
  axElementById("child").takeFocus();
  // Wait for the next frame for the ax object to be recreated.
  requestAnimationFrame(() => {
    requestAnimationFrame(() => {
      axHidden = axElementById("hidden");
      // #hidden is now unlocked and saved as a normal AXLayoutObject with 2 child nodes:
      // "foo" text and #child node.
      t.step(() => { assert_equals(axHidden.childrenCount, 2, "Children count after activation"); });
      t.done();
    });
  });
}, "Accessiblility focus causes activatable hidden tree to activate");
</script>
